[Refactor] stringResource 활용 && ViewModel 활용 && screen 파일 분리
