+++ app/src/main/res/values/strings.xml
@@ -20,4 +20,7 @@
     <string name="pw_text">비밀번호</string>
     <string name="find_id_text">아이디 찾기</string>
     <string name="find_pw_text">비밀번호 찾기</string>
+
+    <string name="sign_up_id_error_msg">아이디는 영문 소문자 또는 영문 소문자, 숫자 조합 6~12자리입니다.</string>
+    <string name="sign_up_pw_error_msg">비밀번호는 영문, 숫자, 특수문자(~!@#$%^&amp;*) 조합 8~15자리입니다.</string>
 </resources>
